[raspbian] Created seprate files for each converter
