+++ netjsonconfig/backends/raspbian/converters/base.py
@@ -1,0 +1,6 @@
+from ...base.converter import BaseConverter
+
+
+class RaspbianConverter(BaseConverter):
+    def test_function(self):
+        return 'Nothing'
