+++ netjsonconfig/backends/raspbian/converters/dnssearch.py
@@ -1,0 +1,13 @@
+from ....utils import get_copy
+from .base import RaspbianConverter
+
+
+class DnsSearch(RaspbianConverter):
+    netjson_key = 'dns_search'
+
+    def to_intermediate(self):
+        result = []
+        dns_search = get_copy(self.netjson, self.netjson_key)
+        for domain in dns_search:
+            result.append(domain)
+        return (('dns_search', result),)
