+++ netjsonconfig/backends/raspbian/raspbian.py
@@ -1,5 +1,5 @@
-from .converters import General, Interfaces, Wireless, DnsServers, DnsSearch, Ntp
-from .renderers import Raspbian
+from . import converters
+from .renderer import Raspbian
 from ..base.backend import BaseBackend
 from .schema import schema
 
@@ -11,11 +11,11 @@ class Raspbian(BaseBackend):
     schema = schema
     env_path = 'netjsonconfig.backends.raspbian'
     converters = [
-        General,
-        Interfaces,
-        Wireless,
-        DnsServers,
-        DnsSearch,
-        Ntp
+        converters.General,
+        converters.Interfaces,
+        converters.Wireless,
+        converters.DnsServers,
+        converters.DnsSearch,
+        converters.Ntp
     ]
     renderer = Raspbian
